requirements.txt
google-cloud-translate==3.3.2